add engine + evaluator. Remove validate from Trainer
