+++ ignite/handlers/__init__.py
@@ -1,1 +1,1 @@
-from .validate import Validate
+from .evaluate import Evaluate
