+++ ignite/handlers/evaluate.py
@@ -1,17 +1,19 @@
-class Validate(object):
+class Evaluate(object):
     """
     Triggers validation and optionally clears history.
 
     Args:
-        validation_data (iterable): the data to validate on
+        evaluator (Evaluator): the evaluator to run
+        evaluation_data (iterable): the data to validate on
         iteration_interval (int, optional): number of iterations between validations
         epoch_interval (int, optional): number of epochs between validations
         clear_history (bool, optional): whether or not to clear history before each
             validation (default: True)
     """
-    def __init__(self, validation_data, iteration_interval=None, epoch_interval=None,
+    def __init__(self, evaluator, evaluation_data, iteration_interval=None, epoch_interval=None,
                  clear_history=True):
-        self._validation_data = validation_data
+        self._evaluator = evaluator
+        self._evaluation_data = evaluation_data
         self._iteration_interval = iteration_interval
         self._epoch_interval = epoch_interval
         self._clear_history = clear_history
@@ -22,12 +24,12 @@ class Validate(object):
             raise ValueError('you must pass one of (iteration_interval, epoch_interval)')
 
     def __call__(self, trainer):
-        if self._should_validate(trainer):
+        if self._should_evaluate(trainer):
             if self._clear_history:
-                trainer.validation_history.clear()
-            trainer.validate(self._validation_data)
+                self._evaluator.history.clear()
+            self._evaluator.run(self._validation_data)
 
-    def _should_validate(self, trainer):
+    def _should_evaluate(self, trainer):
         if self._iteration_interval:
             return (trainer.current_iteration % self._iteration_interval) == 0
         if self._epoch_interval:
