--- ignite/trainer/__init__.py
@@ -1,2 +1,0 @@
-from .trainer import *
-from .history import History
